🛡️ Sentinel: [security improvement] Hardened session and correlation ID generation
- Implemented centralized `generateSecureId` utility in `src/lib/id-utils.ts` to prevent circular dependencies.
- Hardened ID generation across Analytics, Logger, Cloudflare, EventBus, ExportConnectors, ABTest, and TimelineConfig.
- Replaced insecure `Math.random()` with cryptographically secure alternatives for all sensitive identifiers.
- Replaced insecure `console.log` of analytics events with `logger.debug` in `session-analytics.ts`.
- Added unit tests for ID generation and verified Cloudflare compatibility.
- Verified absence of circular dependencies using madge.
